# Core-java_OpenAPI3.yaml
openapi: 3.0.1
info:
  title: vap-core
  description: 'Сервис базового функционала АС ''''Платформа видео-аналитики'''' '
  version: 0.1.0
servers:
  - url: /
...
paths:
  /cameras:
    put:
      tags:
        - cameras
      summary: Сохранить камеру
      description: Сохранить изменения выбранной камеры в БД
      operationId: updateCamera
      requestBody:
        description: Объект типа Camera который нужно добавить в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/camera'
          application/xml:
            schema:
              $ref: '#/components/schemas/camera'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/camera'
            application/json:
              schema:
                $ref: '#/components/schemas/camera'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: Camera not found
          content: {}
        405:
          description: Invalid camera
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - cameras
      summary: Добавить камеру
      description: Добавить новую камеру в БД
      operationId: cameraList
      requestBody:
        description: Объект типа Camera который нужно добавить в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/camera'
          application/xml:
            schema:
              $ref: '#/components/schemas/camera'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/camera'
            application/json:
              schema:
                $ref: '#/components/schemas/camera'
        405:
          description: Invalid camera
          content: {}
      x-codegen-request-body-name: body
  /cameras/{cameraId}:
    get:
      tags:
        - cameras
      summary: Получить камеру
      description: Возвращает камеру с указанным id
      operationId: getCameraById
      parameters:
        - name: cameraId
          in: path
          description: id камеры, которую нужно получить из БД
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/camera'
            application/json:
              schema:
                $ref: '#/components/schemas/camera'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: Camera not found
          content: {}
    delete:
      tags:
        - cameras
      summary: Удалить камеру
      description: Удалить выбранную камеру из БД
      operationId: deleteCamera
      parameters:
        - name: cameraId
          in: path
          description: id камеры, которую нужно удалить
          required: true
          schema:
            minimum: 1.0
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/camera'
            application/json:
              schema:
                $ref: '#/components/schemas/camera'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: Camera not found
          content: {}
...
  /cameras/{cameraId}/nnModels/{nnMmodelId}/:
    get:
      tags:
        - cameras
      summary: Получить видеопоток с камеры
      description: Возвращает видеопоток с выбранной камеры проанализированный выбранной
        нейросетевой моделью
      operationId: getCameraStream
      parameters:
        - name: cameraId
          in: path
          description: id камеры
          required: true
          schema:
            type: integer
            format: int64
        - name: nnMmodelId
          in: path
          description: id нейросетевой модели
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/camera'
            application/json:
              schema:
                $ref: '#/components/schemas/camera'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: Camera or nnModel not found
          content: {}
  /bankUnits:
    get:
      tags:
        - bankUnits
      summary: Список всех подразделений банка
      description: Возвращает список всех подразделений банка из БД
      operationId: getBankUnits
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/bankUnit'
        404:
          description: BankUnits not found
          content: {}
    put:
      tags:
        - bankUnits
      summary: Сохранить подразделения банка
      description: Сохранить изменения выбранного подразделения банка в БД
      operationId: updateBankUnit
      requestBody:
        description: Объект типа BankUnit, который нужно добавить в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bankUnit'
          application/xml:
            schema:
              $ref: '#/components/schemas/bankUnit'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/bankUnit'
            application/json:
              schema:
                $ref: '#/components/schemas/bankUnit'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: BankUnit not found
          content: {}
        405:
          description: Invalid bankUnit
          content: {}
      x-codegen-request-body-name: body
    post:
      tags:
        - bankUnits
      summary: Добавить подразделение банка
      description: Добавить новое подразделение банка в БД
      operationId: placeBankUnit
      requestBody:
        description: Объект типа BankUnit, который нужно добавить в БД
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/bankUnit'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/bankUnit'
            application/json:
              schema:
                $ref: '#/components/schemas/bankUnit'
        405:
          description: Invalid bankUnit
          content: {}
      x-codegen-request-body-name: body
  /bankUnits/{bankUnitId}:
    delete:
      tags:
        - bankUnits
      summary: Удалить подразделение банка
      description: Удалить выбранное подразделение банка из БД
      operationId: deleteBankUnit
      parameters:
        - name: bankUnitId
          in: path
          description: id подразделения банка, которое нужно удалить
          required: true
          schema:
            minimum: 1.0
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/bankUnit'
            application/json:
              schema:
                $ref: '#/components/schemas/bankUnit'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: BankUnit not found
          content: {}
  /parentBankUnit/{parentId}/bankUnits/:
    get:
      tags:
        - bankUnits
      summary: Получить список подразделения банка дочерних для выбранного
      description: Возвращает список камер подразделения банка входящих в структуру
        подразделения с parentId
      operationId: getBankUnitByParentId
      parameters:
        - name: parentId
          in: path
          description: id родительского подразделения банка
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/bankUnit'
            application/json:
              schema:
                $ref: '#/components/schemas/bankUnit'
        400:
          description: Invalid id supplied
          content: {}
        404:
          description: BankUnit not found
          content: {}
...
components:
  schemas:
    camera:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bank_unit_id:
          type: integer
          format: int64
        access_login:
          type: string
        access_pwd:
          type: string
        ip_address:
          type: string
        model:
          type: string
        vendor_name:
          type: string
        description:
          type: string
        comment:
          type: string
      xml:
        name: camera
    bankUnit:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: string
        name:
          type: string
        address:
          type: string
        phone_number:
          type: string
        description:
          type: string
        comment:
          type: string
      xml:
        name: bankUnit
...